Probnet

Description

Authors : Vicent Nos Ripolles (Main Author) Enrique Santos (Refactor Code)

Documentation

percents :: RealFrac a => [a] -> [a]

Ratios between consecutive elements (logarithmic differences).

predict1 :: RealFrac a => [a] -> a

This is to assume that the next ratio is close to that of the element with the closest value to the last element; in case of monotonic data (always increasing or always decreasing) it is the last ratio.

predict :: (Integral b, RealFrac a) => Int -> [a] -> [b]

Generates new prediction

probnet :: (Integral b, RealFrac a) => Int -> [a] -> [b]

Generate new prediction with error prediction